750694: halfmoon does not compile on windows (p=wmaddox, r=fklockii, r=throdrig)
CL@1096510

+++ halfmoon/hm-deoptimizer.cpp
@@ -519,7 +519,7 @@ HMDeoptContext::~HMDeoptContext()
 
 uint8_t HMDeoptContext::readUInt8()
 {
-  return *(mdins_++);
+  return uint8_t(*(mdins_++));
 }
 
 uint32_t HMDeoptContext::readUInt32()
